+++ docs/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Fixed
+- Taskman throwing away logs upon JSON decode error (OSIDB-3296)
+
 ## [4.1.7] - 2024-08-22
 ### Added
 - Command for manual syncing Jira metadata (OSIDB-3219)